Move proofs ecto module to a separate file
